Moved funder's signature_buff.rs into proto.
