+++ components/funder/src/state.rs
@@ -7,12 +7,12 @@ use num_traits::identities::Zero;
 use crypto::identity::PublicKey;
 use crypto::uid::Uid;
 
-use proto::funder::messages::Ratio;
+use proto::funder::messages::{Ratio, SendFundsReceipt};
 
 use utils::int_convert::usize_to_u64;
 
 use crate::friend::{FriendState, FriendMutation};
-use crate::types::{SendFundsReceipt, AddFriend};
+use crate::types::AddFriend;
 
 
 #[derive(Clone, Serialize, Deserialize)]
