+++ components/proto/src/funder/messages.rs
@@ -129,6 +129,31 @@ pub enum FriendMessage {
     InconsistencyError(ResetTerms),
 }
 
+/// A `SendFundsReceipt` is received if a `RequestSendFunds` is successful.
+/// It can be used a proof of payment for a specific `invoice_id`.
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct SendFundsReceipt {
+    pub response_hash: HashResult,
+    // = sha512/256(requestId || sha512/256(route) || randNonce)
+    pub invoice_id: InvoiceId,
+    pub dest_payment: u128,
+    pub signature: Signature,
+    // Signature{key=recipientKey}(
+    //   "FUND_SUCCESS" ||
+    //   sha512/256(requestId || sha512/256(route) || randNonce) ||
+    //   invoiceId ||
+    //   destPayment
+    // )
+}
+
+#[derive(Eq, PartialEq, Debug, Clone, Serialize, Deserialize)]
+pub struct PendingFriendRequest {
+    pub request_id: Uid,
+    pub route: FriendsRoute,
+    pub dest_payment: u128,
+    pub invoice_id: InvoiceId,
+}
+
 // ==================================================================
 // ==================================================================
 
@@ -302,5 +327,16 @@ impl FriendsRoute {
     /// Get the public key of a node according to its index.
     pub fn index_to_pk(&self, index: usize) -> Option<&PublicKey> {
         self.public_keys.get(index)
+    }
+}
+
+impl SendFundsReceipt {
+    pub fn to_bytes(&self) -> Vec<u8> {
+        let mut res_bytes = Vec::new();
+        res_bytes.extend_from_slice(&self.response_hash);
+        res_bytes.extend_from_slice(&self.invoice_id);
+        res_bytes.write_u128::<BigEndian>(self.dest_payment).unwrap();
+        res_bytes.extend_from_slice(&self.signature);
+        res_bytes
     }
 }
