+++ components/proto/src/funder/mod.rs
@@ -1,2 +1,3 @@
 pub mod messages;
 pub mod consts;
+pub mod signature_buff;
